+++ NEWS.md
@@ -7,7 +7,7 @@
   of its values and, conversely, to rebuild an object of type `T` from its
   values.
 
-- New methods `struct_length` yields the number of values needed to destructure
+- New method `struct_length` yields the number of values needed to destructure
   an object.
 
 # Version 0.3.3